(01/12/01) Erwann Chenede - <erwann.chenede@sun.com> * merged changes from HEAD into gtk-multihead branch. * fixed resulting bugs. * added 3 new sgml doc files in docs/reference/gdk/tmpl

+++ gdk/x11/gdkmain-x11.c
@@ -122,7 +122,7 @@ _gdk_windowing_init_check_for_display (int argc,
 {
   XKeyboardState keyboard_state;
   XClassHint *class_hint;
-  guint pid;
+  gulong pid;
   GdkDisplay *display;
   GdkDisplayImplX11 *display_impl;
